Crate cargo_reg_local

cargo_reg_local

Find data from crates.io registry index in local cache.
For now on linux only. The folder of the cache is this: ~\.cargo\registry\index\github.com-1ecc6299db9ec823\.cache\

The only argument is a crate name or a substring of the crate name.

The CLI returns:

a list of versions for a given crate_name

all the crate_names (and last version) that contain the given substring

Build and run

clear; cargo make dev

and then use the example how run it in the last 4th line. Like this:

target/debug/cargo_reg_local thread

Functions

get_versions
Get versions from the local cache of the crates.io index.

main
The CLI program starts here.
Linux only because of file paths. Could be upgraded.
One argument: crate name or a substring of crate name.

search_crates
Search local crates.io index cache for a crate name that contains a substring.

search_file_name_recursive
Search for files_names containing a substring, recursive.